Убрал понятие сцены и сделал много экспортных переменных

+++ xrpg/crt.h
@@ -14,7 +14,7 @@ typedef decltype(sizeof(0)) size_t;
 #define	BSDATATMPL template<>
 #endif
 
-#define maptbl(t, id) (t[imax((unsigned)0, imin((unsigned)id, (sizeof(t)/sizeof(t[0])-1)))])
+#define maptbl(t, id) (t[imax((unsigned long)0, imin((unsigned long)id, (unsigned long)(sizeof(t)/sizeof(t[0])-1)))])
 #define maprnd(t) t[rand()%(sizeof(t)/sizeof(t[0]))]
 #define	FO(T,R) ((size_t)&((T*)0)->R)
 #define BSDATA(e) BSDATATMPL e bsdata<e>::elements[]
